# swagger.yaml
basePath: /api
consumes:
    - application/json
definitions:
    BookReq:
        properties:
            Author:
                type: string
            PublicationYear:
                type: string
            Title:
                type: string
        type: object
        x-go-package: bookify/infra/serializers
    BookResp:
        properties:
            Author:
                type: string
            ID:
                format: uint64
                type: integer
            PublicationYear:
                type: string
            Title:
                type: string
            UpdatedAt:
                format: date-time
                type: string
        type: object
        x-go-package: bookify/infra/serializers
    RestErr:
        properties:
            error:
                example: bad_request
                type: string
                x-go-name: Error
            message:
                example: error message
                type: string
                x-go-name: Message
            status:
                example: 400
                format: int64
                type: integer
                x-go-name: Status
        type: object
        x-go-package: bookify/infra/errors
host: localhost:8080
info:
    contact:
        email: rezwanul.cse@gmail.com
        name: Rezwanul-Haque
    description: the purpose of this service is to provides APIs related to book service
    license:
        name: None
    title: bookify system API.
    version: v1.0.0
paths:
    /v1/book:
        post:
            description: Create a book
            operationId: Post
            responses:
                "200":
                    $ref: '#/responses/BookCreatedResponse'
                "400":
                    $ref: '#/responses/errorResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - Book
        get:
            description: fetch all books
            operationId: List
            responses:
                "200":
                    $ref: '#/responses/BookResponse'
                "400":
                    $ref: '#/responses/errorResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - Book
    /v1/book/{id}:
        delete:
            description: Delete a book
            operationId: Delete
            parameters:
                - name: id
                  in: path
                  description: ID of the book to delete
                  required: true
                  type: string
            responses:
                "204":
                    description: ""
                "400":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - Book
        put:
            description: Update a book
            operationId: Update
            parameters:
                - name: id
                  in: path
                  description: ID of the book to update
                  required: true
                  type: string
            responses:
                "200":
                    $ref: '#/responses/genericSuccessResponse'
                "400":
                    $ref: '#/responses/errorResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - Book
        get:
            description: fetch a book
            operationId: Get
            parameters:
                - name: id
                  in: path
                  description: ID of the book to fetch
                  required: true
                  type: string
            responses:
                "200":
                    $ref: '#/responses/BookCreatedResponse'
                "400":
                    $ref: '#/responses/errorResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - Book
produces:
    - application/json
responses:
    BookCreatedResponse:
        description: response after a book created
        schema:
            $ref: '#/definitions/BookResp'
    BookResponse:
        description: List all the books
        schema:
            items:
                $ref: '#/definitions/BookResp'
            type: array
    errorResponse:
        description: Generic error message
        schema:
            $ref: '#/definitions/RestErr'
    genericSuccessResponse:
        description: returns a message
        headers:
            message:
                example: resource created
                type: string
schemes:
    - http
swagger: "2.0"